refactor: rename data/ -> memory/ to headline the Lakebase agent-memory story
Pure refactor — no behavior change. Uses git mv to preserve history.
Updates all imports: app.py (from data import layer), memory/lakebase.py
(from data.credentials -> from memory.credentials). Adds conftest.py so
local packages are on sys.path during pytest, and adds tests/test_imports.py
(TDD green) verifying memory.lakebase, memory.credentials, memory.layer.
