+++ src/app/app.py
@@ -1,6 +1,6 @@
 import chainlit as cl
 import logging
 from typing import Optional, Dict
-from data import layer
+from memory import layer
 from auth import header, password_auth
 import routes
